+++ src/synaudio-cli.js
@@ -1,251 +1,8 @@
 import SynAudio from "synaudio";
 import fs from "fs";
-import { spawn } from "child_process";
-
-const copyToSharedMemoryChunked = (
-  inputChunks,
-  inputLength,
-  outputChunkSeconds,
-  outputChunkIntervalSeconds,
-  syncStartRange,
-  syncEndRange,
-  sampleRate,
-) => {
-  const sharedMemories = [];
-  const samplesPerChunk = sampleRate * outputChunkSeconds;
-  const samplesToSkip = sampleRate * outputChunkIntervalSeconds;
-
-  let sharedMemory = null;
-
-  let currentPosition = 0; // total bytes read
-  let nextStart = 0;
-  let nextEnd = 0;
-  let sharedMemoryPosition = 0;
-
-  for (const chunk of inputChunks) {
-    if (currentPosition >= nextEnd) {
-      // start a new output chunk
-      nextStart = nextEnd + samplesToSkip;
-      nextEnd = Math.min(nextStart + samplesPerChunk, inputLength);
-
-      if (nextStart >= inputLength) {
-        // next output chunk would start be beyond the input data
-        break;
-      }
-      const sharedMemoryBuffer = new SharedArrayBuffer(
-        (nextEnd - nextStart) * Float32Array.BYTES_PER_ELEMENT,
-      );
-      sharedMemory = new Float32Array(sharedMemoryBuffer);
-      sharedMemories.push({
-        start: nextStart,
-        end: nextEnd,
-        syncStart:
-          Math.round(
-            sharedMemories.length * outputChunkIntervalSeconds - syncStartRange,
-          ) * sampleRate,
-        syncEnd:
-          Math.round(
-            sharedMemories.length * outputChunkIntervalSeconds + syncEndRange,
-          ) * sampleRate,
-        data: sharedMemory,
-      });
-      sharedMemoryPosition = 0;
-    }
-
-    const chunkStart = nextStart - currentPosition;
-    const chunkEnd = nextEnd - currentPosition;
-
-    if (chunkStart < chunk.length) {
-      const chunkLength = chunk.length;
-      const startOffset = Math.max(0, chunkStart);
-      const endOffset = Math.min(chunkLength, chunkEnd);
-      const chunkToCopy = chunk.subarray(startOffset, endOffset);
-      sharedMemory.set(chunkToCopy, sharedMemoryPosition);
-      sharedMemoryPosition += chunkToCopy.length;
-    }
-
-    currentPosition += chunk.length;
-  }
-
-  return sharedMemories;
-};
-
-const copyToSharedMemory = (inputChunks, inputLength) => {
-  let currentPosition = 0;
-  const sharedMemoryBuffer = new SharedArrayBuffer(
-    inputLength * Float32Array.BYTES_PER_ELEMENT,
-  );
-  const sharedMemory = new Float32Array(sharedMemoryBuffer);
-
-  for (const chunk of inputChunks) {
-    sharedMemory.set(chunk, currentPosition);
-    currentPosition += chunk.length;
-  }
-
-  return sharedMemory;
-};
-
-const decodeAndSave = async (
-  inputFile,
-  sampleRate,
-  rectify,
-  chunkSize = 0,
-  chunkInterval = 0,
-  syncStartRange,
-  syncEndRange,
-) => {
-  const { promise, stdout } = runCmd("ffmpeg", [
-    "-v",
-    "error",
-    "-i",
-    inputFile,
-    "-map",
-    "0:a:0", // only support one audio file
-    "-ac",
-    "1",
-    "-ar",
-    sampleRate,
-    "-f",
-    "f32le",
-    "-acodec",
-    "pcm_f32le",
-    "-",
-  ]);
-
-  const outputChunks = [];
-  let outputSamples = 0;
-
-  stdout.on("data", (data) => {
-    const decoded = new Float32Array(data.buffer);
-
-    // rectify to avoid comparison issues with phase
-    if (rectify) {
-      for (let i = 0; i < decoded.length; i++) {
-        decoded[i] = decoded[i] < 0 ? -decoded[i] : decoded[i];
-      }
-    }
-    outputChunks.push(decoded);
-    outputSamples += decoded.length;
-  });
-
-  await promise;
-
-  return chunkSize === 0
-    ? copyToSharedMemory(outputChunks, outputSamples)
-    : copyToSharedMemoryChunked(
-        outputChunks,
-        outputSamples,
-        chunkSize,
-        chunkInterval,
-        syncStartRange,
-        syncEndRange,
-        sampleRate,
-      );
-};
-
-function removeOutliers(data, tolerance = 0.5) {
-  // Step 1: Build frequency map of rounded differences
-  const freqMap = {};
-
-  for (const item of data) {
-    const rounded = Math.round(item.difference * 10) / 10;
-    freqMap[rounded] = (freqMap[rounded] || 0) + 1;
-  }
-
-  // Step 2: Find the mode (most common rounded value)
-  let mode = null;
-  let maxCount = -1;
-
-  for (const [valueStr, count] of Object.entries(freqMap)) {
-    const value = parseFloat(valueStr);
-    if (count > maxCount) {
-      maxCount = count;
-      mode = value;
-    }
-  }
-
-  // Step 3: Define bounds around the mode (±0.5)
-  const lowerBound = mode - tolerance;
-  const upperBound = mode + tolerance;
-
-  // Step 4: Split data into good values and outliers
-  const goodValues = data.filter(
-    (item) => item.difference >= lowerBound && item.difference <= upperBound,
-  );
-  const outliers = data.filter(
-    (item) => item.difference < lowerBound || item.difference > upperBound,
-  );
-  return goodValues;
-}
-
-function weightedLinearRegression(data) {
-  const cleaned = removeOutliers(data);
-  let sumW = 0,
-    sumWX = 0,
-    sumWY = 0,
-    sumWXY = 0,
-    sumWXX = 0;
-
-  for (const d of cleaned) {
-    const w = d.correlation;
-    const x = d.order;
-    const y = d.difference;
-
-    sumW += w;
-    sumWX += w * x;
-    sumWY += w * y;
-    sumWXY += w * x * y;
-    sumWXX += w * x * x;
-  }
-
-  const denominator = sumW * sumWXX - sumWX * sumWX;
-
-  const slope = (sumW * sumWXY - sumWX * sumWY) / denominator;
-  const intercept = (sumWY * sumWXX - sumWX * sumWXY) / denominator;
-
-  return { slope, intercept };
-}
-
-function simpleLinearRegression(data, maxRateTolerance) {
-  const cleaned = removeOutliers(data, maxRateTolerance);
-  const n = cleaned.length;
-  let sumX = 0,
-    sumY = 0,
-    sumXY = 0,
-    sumX2 = 0;
-
-  for (const point of cleaned) {
-    const x = point.order;
-    const y = point.difference;
-    sumX += x;
-    sumY += y;
-    sumXY += x * y;
-    sumX2 += x * x;
-  }
-
-  const slope = (n * sumXY - sumX * sumY) / (n * sumX2 - sumX * sumX);
-  const intercept = (sumY - slope * sumX) / n;
-
-  return { slope, intercept };
-}
-
-const runCmd = (cmd, args) => {
-  const spawned = spawn(cmd, args);
-  const promise = new Promise((resolve, reject) => {
-    spawned.stderr.on("data", (data) => {
-      process.stderr.write(data);
-    });
-
-    spawned.on("close", (code) => {
-      if (code !== 0) {
-        console.error(`${cmd} ${args.join(" ")} exit code ${code}`);
-      }
-      resolve(code);
-    });
-  });
-
-  return { promise, stdin: spawned.stdin, stdout: spawned.stdout };
-};
+import { decodeAndSave } from "./decode.js";
+import { simpleLinearRegression } from "./linear-regression.js";
+import { getFileInfo, runCmd } from "./utilities.js";
 
 const trimAndResample = async (
   inputFile,
@@ -342,6 +99,7 @@ const trimAndResample = async (
       "-",
     ]);
     const { promise: flacPromise, stdin: flacStdin } = runCmd("flac", [
+      "-s",
       "--endian=little",
       "--sign=signed",
       "--bps=" + bitDepth,
@@ -363,47 +121,6 @@ const trimAndResample = async (
     await Promise.all(tempFiles.map((file) => fs.promises.rm(file).catch()));
   }
 };
-
-const getFileInfo = async (files) =>
-  Promise.all(
-    files.map(
-      (file) =>
-        new Promise((resolve, reject) => {
-          const buffers = [];
-          const { promise, stdout } = runCmd("ffprobe", [
-            "-v",
-            "error",
-            "-select_streams",
-            "a",
-            "-show_entries",
-            "stream=channels,bits_per_raw_sample,sample_rate",
-            "-of",
-            "json",
-            file,
-          ]);
-
-          stdout.on("data", (data) => {
-            buffers.push(data);
-          });
-
-          promise.then(() => {
-            const fileInfo = JSON.parse(
-              Buffer.concat(buffers).toString("utf8"),
-            );
-            const firstStream = fileInfo?.streams?.[0]; // only supports using the first stream
-            if (!firstStream) {
-              reject(file + " does not have any audio streams");
-            } else {
-              resolve({
-                sampleRate: firstStream["sample_rate"],
-                channels: firstStream["channels"],
-                bitDepth: parseInt(firstStream["bits_per_raw_sample"]),
-              });
-            }
-          });
-        }),
-    ),
-  );
 
 export const syncResampleEncode = async ({
   baseFile,
